base: use database_host to figure out the connection class
In a multiple-databse setup like such:

  DatabaseCleaner[:active_record, db: :foo].strategy = :transaction
  DatabaseCleaner[:active_record, db: :bar].strategy = :transaction

the connection hash will usually be computer properly but the
`lookup_from_connection_pool` method is faulty because it tries to
find the right ActiveRecord::Base class based on the database name,
which might be the same for both databases (they were both called
`test` in my case, which is fair).

Checking against the host does a better job of identifying the right
model, though I'm realising that you might have two databases on the
same host. TBC.